requirements.txt
bottle
google-cloud-datastore
google-cloud-logging